Flattening Transport into ClientTransport and ServerTransport
