+++ src/replit_river/server.py
@@ -10,7 +10,7 @@ from replit_river.messages import WebsocketClosedException
 from replit_river.seq_manager import SessionStateMismatchException
 from replit_river.server_session import ServerSession
 from replit_river.server_transport import ServerTransport
-from replit_river.transport import TransportOptions
+from replit_river.transport_options import TransportOptions
 
 from .rpc import (
     GenericRpcHandler,
