+++ src/replit_river/server_transport.py
@@ -1,7 +1,9 @@
+import asyncio
 import logging
-from typing import Any, Mapping
+from typing import Any, Callable, Mapping
 
 import nanoid  # type: ignore  # type: ignore
+from grpc import GenericRpcHandler
 from pydantic import ValidationError
 from websockets import (
     WebSocketCommonProtocol,
@@ -29,26 +31,43 @@ from replit_river.seq_manager import (
 )
 from replit_river.server_session import ServerSession
 from replit_river.session import Session
-from replit_river.transport import Transport
 from replit_river.transport_options import TransportOptions
 
 logger = logging.getLogger(__name__)
 
 
-class ServerTransport(Transport):
+class ServerTransport:
     _sessions: dict[str, ServerSession]
+    _handlers: dict[tuple[str, str], tuple[str, GenericRpcHandler]]
 
     def __init__(
         self,
         transport_id: str,
         transport_options: TransportOptions,
     ) -> None:
-        super().__init__(
-            transport_id=transport_id,
-            transport_options=transport_options,
-            is_server=True,
-        )
         self._sessions = {}
+        self._transport_id = transport_id
+        self._transport_options = transport_options
+        self._handlers: dict[tuple[str, str], tuple[str, GenericRpcHandler]] = {}
+        self._session_lock = asyncio.Lock()
+
+    async def _close_all_sessions(
+        self,
+        get_all_sessions: Callable[[], Mapping[str, Session]],
+    ) -> None:
+        sessions = get_all_sessions().values()
+        logger.info(
+            f"start closing sessions {self._transport_id}, number sessions : "
+            f"{len(sessions)}"
+        )
+        sessions_to_close = list(sessions)
+
+        # closing sessions requires access to the session lock, so we need to close
+        # them one by one to be safe
+        for session in sessions_to_close:
+            await session.close()
+
+        logger.info(f"Transport closed {self._transport_id}")
 
     async def handshake_to_get_session(
         self,
@@ -114,7 +133,7 @@ class ServerTransport(Transport):
                     session_id,
                     websocket,
                     self._transport_options,
-                    self._is_server,
+                    True,
                     self._handlers,
                     close_session_callback=self._delete_session,
                 )
@@ -135,7 +154,7 @@ class ServerTransport(Transport):
                         session_id,
                         websocket,
                         self._transport_options,
-                        self._is_server,
+                        True,
                         self._handlers,
                         close_session_callback=self._delete_session,
                     )
